Project xử lí hình ảnh/backend/requirements.txt
Flask
opencv-python
numpy
Pillow
rembg
flask-cors
